Moved the StudentQuestions to matching app and split up the tests into their respective apps
