--- api/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-from .models import QuestionSlider
-
-admin.site.register(QuestionSlider)
